Assistant checkpoint: Add iOS home screen icon support
Assistant generated file changes:
- public/index.html: Add iOS home screen icon meta tags

---

User prompt:

On the Pixel Android, When I add the website link as a Home Screen icon, it automatically used the favicon.

However, on Apple iPhone, when I added the website link as a Home Screen icon, it did not pick up the favicon.

Please fix this.