Assistant checkpoint: Add nearest international day functionality
Assistant generated file changes:
- server.js: Add function to get nearest international day
- public/js/main.js: Update inspiration modal to show nearest international day

---

User prompt:

Write up a function to do the following:
Given the current time, using the month and date of the month, we need to lookup of the Column A in   "InternationalDays.csv" file. Find the next nearest Date in Column A and return the corresponding date, the corresponding theme and the URL (if it is available).

Call this function when the "Need Inspiration" link is clicked and display the details in the UI.